+++ module-system/module-implementations/sov-accounts/src/hooks.rs
@@ -1,33 +1,44 @@
 use sov_modules_api::transaction::AuthenticatedTransactionData;
-use sov_modules_api::{Hash, Spec, StateAccessor, StateCheckpoint};
+use sov_modules_api::{Spec, StateCheckpoint};
 
 use crate::{Account, Accounts};
 
 impl<S: Spec> Accounts<S> {
-    fn get_or_create_default(
+    /// Resolve the sender's public key to an address. Return an error if the sender is not registered.
+    pub fn resolve_sender_address(
         &self,
-        pub_key_hash: &Hash,
-        default_address: &S::Address,
-        working_set: &mut impl StateAccessor,
-    ) -> Account<S> where {
-        if let Some(acct) = self.accounts.get(pub_key_hash, working_set) {
-            acct
-        } else {
-            let new_account = Account {
-                addr: default_address.clone(),
-                nonce: 0,
-            };
+        tx: &AuthenticatedTransactionData<S>,
+        state_checkpoint: &mut StateCheckpoint<S>,
+    ) -> Result<S::Address, anyhow::Error> {
+        let pub_key_hash = &tx.pub_key_hash;
+        let maybe_address = self
+            .accounts
+            .get(pub_key_hash, state_checkpoint)
+            .map(|a| a.addr);
 
-            self.accounts.set(pub_key_hash, &new_account, working_set);
+        match maybe_address {
+            Some(address) => Ok(address),
+            None => match &tx.default_address {
+                Some(default_address) => {
+                    let new_account = Account {
+                        addr: default_address.clone(),
+                        nonce: 0,
+                    };
 
-            self.public_keys
-                .set(default_address, pub_key_hash, working_set);
-            new_account
+                    self.accounts
+                        .set(pub_key_hash, &new_account, state_checkpoint);
+
+                    self.public_keys
+                        .set(default_address, pub_key_hash, state_checkpoint);
+
+                    Ok(default_address.clone())
+                }
+                None => anyhow::bail!("No default address found for {}", pub_key_hash),
+            },
         }
     }
 
     /// Checks that a transaction is not a duplicate
-    // TODO(@preston-evans98): Enforce that this is read-only
     pub fn check_uniqueness(
         &self,
         tx: &AuthenticatedTransactionData<S>,
@@ -35,12 +46,15 @@ impl<S: Spec> Accounts<S> {
     ) -> Result<(), anyhow::Error> {
         // TODO(@preston-evans98) - this check should rely on the information resolved from the context.
         // This will require a change to the account state layout
+        let pub_key_hash = &tx.pub_key_hash;
         let sender_nonce = self
             .accounts
-            .get(tx.pub_key_hash(), state_checkpoint)
+            .get(pub_key_hash, state_checkpoint)
             .map(|a| a.nonce)
-            .unwrap_or(0);
-        let tx_nonce = tx.nonce();
+            .unwrap_or_else(|| panic!("The existence of the sender account for {} is ensured during the resolution of the sender's address.",
+              &pub_key_hash));
+
+        let tx_nonce = tx.nonce;
 
         anyhow::ensure!(
             sender_nonce == tx_nonce,
@@ -57,25 +71,16 @@ impl<S: Spec> Accounts<S> {
         tx: &AuthenticatedTransactionData<S>,
         state_checkpoint: &mut StateCheckpoint<S>,
     ) {
-        // TODO(@preston-evans98) - this check should rely on the information resolved from the context.
-        // This will require a change to the account state layout
-        let mut account =
-            self.get_or_create_default(tx.pub_key_hash(), tx.default_address(), state_checkpoint);
+        let pub_key_hash = &tx.pub_key_hash;
+        let mut account = self
+            .accounts
+            .get(pub_key_hash, state_checkpoint)
+            .unwrap_or_else(|| panic!("The existence of the sender account for {} is ensured during the resolution of the sender's address.",
+                &pub_key_hash));
+
         account.nonce += 1;
 
         self.accounts
-            .set(tx.pub_key_hash(), &account, state_checkpoint);
-    }
-
-    /// Resolve the sender public key to an address
-    pub fn resolve_sender_address(
-        &self,
-        tx: &AuthenticatedTransactionData<S>,
-        state_checkpoint: &mut StateCheckpoint<S>,
-    ) -> S::Address {
-        self.accounts
-            .get(tx.pub_key_hash(), state_checkpoint)
-            .map(|a| a.addr)
-            .unwrap_or(tx.default_address().clone())
+            .set(&tx.pub_key_hash, &account, state_checkpoint);
     }
 }